+++ src/Factory.php
@@ -91,7 +91,7 @@ class Factory
      */
     public function view($view, array $data = [])
     {
-        return with(new Facile($this, 'default'))
+        return (new Facile($this, 'default'))
             ->view($view)
             ->with($data);
     }
@@ -157,8 +157,8 @@ class Factory
      */
     public function getRequestFormat($name)
     {
-        return $this->request->format(
-            $this->getTemplate($name)->getDefaultFormat()
+        return $this->request->prefers(
+            $this->getTemplate($name)->getSupportedFormats()
         );
     }