+++ src/Template/Simple.php
@@ -93,7 +93,7 @@ class Simple extends Template
                 $content = $content->toArray();
             }
 
-            $content = with(new Collection(array_map([$this, 'transformToArray'], $content)));
+            $content = (new Collection(array_map([$this, 'transformToArray'], $content)));
         }
 
         return new IlluminateResponse($content->toCsv(), $status, [